+++ tika-parsers/src/main/java/org/apache/tika/parser/chm/lzx/ChmSection.java
@@ -61,7 +61,7 @@ public class ChmSection {
     }
 
     public int unmarshalUByte() {
-        return (int) (getByte() & 255);
+        return getByte() & 255;
     }
 
     public byte getByte() {
